+++ stateproofs/stateprooftypes/generic_digest.go
@@ -1,4 +1,4 @@
-package datatypes
+package stateprooftypes
 
 import "bytes"
