+++ stateproofs/stateprooftypes/light_block_header.go
@@ -1,4 +1,4 @@
-package datatypes
+package stateprooftypes
 
 import (
 	"github.com/algorand/go-algorand-sdk/encoding/msgpack"
